newsblog/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
drf-spectacular>=0.27,<0.28
django-filter>=24.2
psycopg2-binary>=2.9
Pillow>=10.2
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=21.2
nltk>=3.8
requests>=2.31